# openapi.yaml
openapi: 3.1.0
info:
  title: AgilePlus REST API
  description: |
    MVP OpenAPI scaffold — 5 representative endpoints spanning features,
    work packages, events, audit, and governance. Full coverage of all 36
    handlers is tracked as a follow-up PR.

    This document was hand-authored from the `#[utoipa::path]` annotations
    in `crates/agileplus-api/` because the workspace cannot currently
    `cargo build` — the `agileplus-plugin-core` git dependency (SHA
    `ba652dec…`) returns HTTP 404. Once that dep is unpinned or republished,
    run `cargo run -p agileplus-api --bin agileplus-api -- --dump-openapi \
    > openapi.yaml` and let the CI drift check take over.
  version: 0.1.1
  license:
    name: MIT
...
components:
  securitySchemes:
    api_key:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    GovernanceResponse:
      type: object
      required: [id, feature_id, version, rules_count, bound_at]
      properties:
        id:
          type: integer
          format: int64
        feature_id:
          type: integer
          format: int64
        version:
          type: integer
          format: int32
        rules_count:
          type: integer
          format: int64
        bound_at:
          type: string
          format: date-time
paths:
...
  /api/v1/features/{slug}/governance:
    get:
      tags: [governance]
      summary: Get latest governance contract for a feature
      operationId: get_governance
      security:
        - api_key: []
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
          description: Feature slug
      responses:
        '200':
          description: Latest governance contract for feature
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GovernanceResponse'
        '404':
          description: Feature or contract not found
        '401':
          description: Missing or invalid API key